pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "be-humming"
version = "1.0.0"
description = "A next-generation, perception-aware image compressor toolkit."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Multimedia :: Graphics :: GraphicsConversion",
    "Topic :: Utilities",
]

dependencies = [
    "pillow",
    "tqdm",
    "numpy",
    "scikit-image",
    "opencv-python-headless",
]

[project.scripts]
be-humming = "be_humming.main:main"